+++ src/pipecat/services/gemini_multimodal_live/gemini.py
@@ -229,15 +229,12 @@ class GeminiMultimodalLiveLLMService(LLMService):
     #
 
     async def start(self, frame: StartFrame):
-        await super().start(frame)
         await self._connect()
 
     async def stop(self, frame: EndFrame):
-        await super().stop(frame)
         await self._disconnect()
 
     async def cancel(self, frame: CancelFrame):
-        await super().cancel(frame)
         await self._disconnect()
 
     #