+++ src/pipecat/services/gladia.py
@@ -177,18 +177,15 @@ class GladiaSTTService(STTService):
         return language_to_gladia_language(language)
 
     async def start(self, frame: StartFrame):
-        await super().start(frame)
         response = await self._setup_gladia()
         self._websocket = await websockets.connect(response["url"])
         self._receive_task = self.get_event_loop().create_task(self._receive_task_handler())
 
     async def stop(self, frame: EndFrame):
-        await super().stop(frame)
         await self._send_stop_recording()
         await self._websocket.close()
 
     async def cancel(self, frame: CancelFrame):
-        await super().cancel(frame)
         await self._websocket.close()
 
     async def run_stt(self, audio: bytes) -> AsyncGenerator[Frame, None]: