+++ src/pipecat/services/openai_realtime_beta/openai.py
@@ -112,15 +112,12 @@ class OpenAIRealtimeBetaLLMService(LLMService):
     #
 
     async def start(self, frame: StartFrame):
-        await super().start(frame)
         await self._connect()
 
     async def stop(self, frame: EndFrame):
-        await super().stop(frame)
         await self._disconnect()
 
     async def cancel(self, frame: CancelFrame):
-        await super().cancel(frame)
         await self._disconnect()
 
     #